+++ src/config/system/strategies.ts
@@ -187,6 +187,86 @@ export const getStrategies = (network?: string | null): StrategyNetworkConfig =>
 					strategyLink:
 						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#418b98a05da849a3a8dd97f74f8c0c80',
 				},
+				[deploy.sett_system.vaults['experimental.digg']]: {
+					name: '',
+					address: deploy.sett_system.vaults['experimental.digg'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(1000),
+						['Strategist Performance Fee']: new BigNumber(1000),
+						['Withdraw Fee']: new BigNumber(20),
+					},
+					strategyLink: '',
+				},
+				[deploy.sett_system.vaults['native.hbtcCrv']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.hbtcCrv'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(3000),
+						['Withdraw Fee']: new BigNumber(20),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#56cd7b65cd384740aa9f339bf3ee2597',
+				},
+				[deploy.sett_system.vaults['native.pbtcCrv']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.pbtcCrv'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(3000),
+						['Withdraw Fee']: new BigNumber(20),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#39a0decd933b4869b98c9276118b9d39',
+				},
+				[deploy.sett_system.vaults['native.obtcCrv']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.obtcCrv'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(3000),
+						['Withdraw Fee']: new BigNumber(20),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#82d72e94cb3b49f0836d8197ad13bc36',
+				},
+				[deploy.sett_system.vaults['native.bbtcCrv']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.bbtcCrv'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(3000),
+						['Withdraw Fee']: new BigNumber(20),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#fe4a64edc830472da5a700d0fc30716c',
+				},
+				[deploy.sett_system.vaults['native.tricryptoCrv']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.tricryptoCrv'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(3000),
+						['Withdraw Fee']: new BigNumber(20),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#f03b01a576d241aa9d9cee153876c976',
+				},
+				[deploy.sett_system.vaults['native.cvxCrv']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.cvxCrv'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(1000),
+						['Withdraw Fee']: new BigNumber(10),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#51d48102bc4847a6a5a1a059c4b827b3',
+				},
+				[deploy.sett_system.vaults['native.cvx']]: {
+					name: '',
+					address: deploy.sett_system.vaults['native.cvx'],
+					fees: {
+						['DAO Performance Fee']: new BigNumber(1000),
+						['Withdraw Fee']: new BigNumber(10),
+					},
+					strategyLink:
+						'https://badger.wiki/Strategies-7bf5b27a451242538f02855ca5aaf4e4#1346adfaad7946eebd29a17fb4f6e8b7',
+				},
 			};
 	}
 };